refactoring and bugfixes
